+++ android/src/car/superfun/game/gameModes/gladiatorMode/LocalGladiatorCar.java
@@ -1,10 +1,9 @@
 package car.superfun.game.gameModes.gladiatorMode;
 
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.World;
 
-import car.superfun.game.car.CarController;
+import car.superfun.game.car.LocalCarController;
 import car.superfun.game.car.LocalCar;
 
 import static java.lang.Math.abs;
@@ -15,8 +14,8 @@ public class LocalGladiatorCar extends LocalCar {
     private int score;
     
 
-    public LocalGladiatorCar(Vector2 position, CarController carController, World world, Integer score){
-        super(position, carController, world);
+    public LocalGladiatorCar(Vector2 position, LocalCarController localCarController, World world, Integer score){
+        super(position, localCarController, world);
         this.score = score;
     }
