+++ android/src/car/superfun/game/gameModes/raceMode/LocalRaceCar.java
@@ -6,15 +6,15 @@ import com.badlogic.gdx.physics.box2d.World;
 
 import java.util.Arrays;
 
-import car.superfun.game.car.CarController;
+import car.superfun.game.car.LocalCarController;
 import car.superfun.game.car.LocalCar;
 
 public class LocalRaceCar extends LocalCar {
     boolean[] passedCheckpoints;
     int completedRounds;
 
-    public LocalRaceCar(Vector2 position, CarController carController, World world, int amountOfCheckpoints) {
-        super(position, carController, world);
+    public LocalRaceCar(Vector2 position, LocalCarController localCarController, World world, int amountOfCheckpoints) {
+        super(position, localCarController, world);
         passedCheckpoints = new boolean[amountOfCheckpoints];
         Arrays.fill(passedCheckpoints, Boolean.FALSE);
         completedRounds = 0;
